fix(stella-tools,stella-runtime): make the issue-backend probe a declared port
Constructing a tool registry forked a process and read ambient environment:
`ToolRegistry::new_detected` called `detect_issue_backend`, which spawns
`gh auth status` and reads `LINEAR_API_KEY`, the stored OAuth connections, and
(through `TrackerStore::open_default`) `HOME`.

That defeated `stella-runtime`'s documented guarantee — "nothing in this crate
reads the process environment" — one call below the crate it is scoped to:
`RuntimeBuilder::build` reaches `parts::tool_registry` reaches
`new_detected`, so building a `SessionRuntime` read ambient state no
`RuntimeSpec` field mentioned. Two sessions in one process could therefore
disagree about the host, which is the exact property that crate exists to
prevent. `tests/no_ambient_reads.rs` scans only that crate's own sources, so it
was structurally unable to see this.

Detection is now a port (invariant 1). `RegistryOptions::issue_backend` carries
an `IssueBackendSource` — `None`, a `Declared(IssueBackend)`, or a
`Probe(Arc<dyn IssueBackendProbe>)` — and defaults to `None`, so construction
spawns no process and reads no environment until a caller asks. The ambient
implementation still lives in `stella-tools`, where I/O is allowed; what
changed is that reaching it is a decision at the call site rather than a
property of having called a constructor.

The probe trait is deliberately blocking rather than `async`. The sync and
async constructors have to stay behaviorally aligned, and a blocking method is
the shape both can honour exactly: `new` calls it directly, `new_detected`
routes it through the blocking pool, preserving #64. An `async` trait would
leave the sync path unable to call it at all.

`process_free` isolation gates the *probe* and not a declaration. Isolation is
about subprocess authority, and a credential the host handed us spawns
nothing — `with_backends_and_options` has always accepted an explicit backend
under the same isolation.

## Call sites

`agent::tools::registry_options` is the one place the CLI opts in, so every
session driver that already routes through it is unchanged. `stella tools`'
listing opts in too, via `RegistryOptions::ambient()` — its whole job is to
mirror what a real session on this machine would build.

Deliberately NOT opted in, and a behaviour change worth naming: the witness
tool registries (`candidate_ws/witness_tools.rs`) and two test fixtures
(`durability.rs`, `command_deck/session_clear.rs`). None of them register
issue tools for any purpose, so each was paying a `gh auth status` subprocess
per construction for nothing.

## Witnesses

`crates/stella-runtime/tests/registry_probe_is_declared.rs` — behavioural, not
another needle list, because a textual scan cannot see a constructor one call
below:

- `a_default_spec_never_consults_the_host` — a default spec yields a registry
  with no issue tools. Asserted through the registered surface rather than a
  counter: with no probe installed a counter could not move, so it would prove
  nothing.
- `a_declared_probe_is_the_only_thing_consulted` — the declared probe answers
  exactly once and its backend's tools appear.
- `a_claim_mode_trial_probes_nothing_even_when_one_is_declared` — pins the
  persistence-disabled arm so a later refactor merging the two cannot quietly
  hand a claim-mode trial a host probe.

Plus three unit tests in `stella-tools` covering the default, the declaration,
and the process-free split, using a probe that panics if reached.

All four fail on the old code as compile errors: there was no field to set.

## The god file

The port needs one line in a `RegistryOptions` literal in `media.rs`, which
sat at exactly its 1569-line ceiling. Raising a ceiling to turn a gate green
is a defect rather than a tradeoff, so instead its 1077-line inline
`#[cfg(test)] mod tests` moved to `media/tests.rs` — the remedy AGENTS.md
prescribes, and the precedent `src/media/` already sets. `media.rs` is now 491
lines and leaves the god-file list entirely: AGENTS.md, the crate README, and
the baseline all drop it, and `check-god-files` verifies the three agree.

`registry.rs` stays at its ceiling by construction — the detection branches it
lost pay for the field it gained (2182 -> 2181). `agent.rs` takes no new line
because `RegistryOptions::ambient()` swaps a call for a call.

The same regeneration retires three stale ceilings: pipeline.rs 3616->3530,
command_deck.rs 4647->4621, agent.rs 2269->2266.

## Note for review

`crates/stella-cli/src/agent/tools.rs` is now 1496 lines. It is not
grandfathered, so it fails the gate outright at 1500 — four lines of room, and
the next change there should split it rather than spend them.

`cargo test -p stella-cli` 1439 passed; `-p stella-tools` 728 passed;
`-p stella-runtime` all green. Workspace clippy `-D warnings`, `fmt --check`,
rustdoc `-D warnings`, `check-file-size`, `check-god-files`,
`check-module-reachability`, `check-left-behind` and `check-invariants` clean.

Closes #1596
Refs #1557, #1495
