+++ scripts/file-size-baseline.txt
@@ -15,9 +15,9 @@
 1911 bench/terminal_bench_analysis/tb21_evidence_contract.py
 4334 bench/terminal_bench_analysis/tests/test_tb21_analysis.py
 1659 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
-2269 crates/stella-cli/src/agent.rs
+2266 crates/stella-cli/src/agent.rs
 1751 crates/stella-cli/src/agent/tests.rs
-4647 crates/stella-cli/src/command_deck.rs
+4621 crates/stella-cli/src/command_deck.rs
 1504 crates/stella-cli/src/fleet_cmd.rs
 2129 crates/stella-core/src/bus.rs
 2568 crates/stella-core/src/driver.rs
@@ -26,14 +26,13 @@
 2093 crates/stella-model/src/openai.rs
 1565 crates/stella-model/src/zai.rs
 1895 crates/stella-model/src/zai/tests.rs
-3616 crates/stella-pipeline/src/pipeline.rs
+3530 crates/stella-pipeline/src/pipeline.rs
 2572 crates/stella-pipeline/src/pipeline/tests.rs
 2965 crates/stella-protocol/src/event.rs
 1995 crates/stella-store/src/lib.rs
 2268 crates/stella-store/src/tests.rs
 1916 crates/stella-store/src/usage.rs
-1569 crates/stella-tools/src/media.rs
-2182 crates/stella-tools/src/registry.rs
+2181 crates/stella-tools/src/registry.rs
 1839 crates/stella-tools/src/scripts.rs
 1531 crates/stella-tui/src/deck_render.rs
 4006 crates/stella-tui/src/deck_ui.rs
